Target Android 10 (API level 29)
"

Hello Google Play developer,

This is a reminder that starting November 2, 2020, updates to apps and games on Google Play will be required to target Android 10 (API level 29) or higher. After this date, the Play Console will prevent you from submitting new app bundles and APKs with a targetSdkVersion less than 29. Please note that Wear OS apps are not subject to the API level 29 requirement. This will not impact your existing listing on the Play Store"